+++ ganoverflow-next/src/app/accounts/login/page.tsx
@@ -1,8 +1,8 @@
-import LoginBox from "./LoginComp";
+import LoginBox from "./LoginBox";
 
 const Login = () => {
   return (
-    <div className="my-12 flex flex-col justify-center sm:px-6 lg:px-8">
+    <div className=" flex flex-col justify-center sm:px-6 lg:px-8">
       <LoginBox />
     </div>
   );
