+++ ganoverflow-next/src/app/chat/ChatMain/components/SaveChatModal/handlers.tsx
@@ -1,0 +1,7 @@
+  const checkUserSigned = useSignedCheck();
+    const isUserSignedIn = await checkUserSigned();
+
+    if (!isUserSignedIn) {
+      console.log("로그인이 필요합니다.");
+      return;
+    }
