+++ ganoverflow-next/src/app/chat/ChatMain/index.tsx
@@ -1,0 +1,8 @@
+      {isSigned === TIsSigned.F && <LoginBoxModal />}
+      {isModalOpen && (
+        <SaveChatModal
+          authData={authData}
+          categories={categories}
+          setIsModalOpen={setIsModalOpen}
+        />
+      )}
