+++ ganoverflow-next/src/app/chat/ChatSideBar/components/SideBarHeader/handlers.tsx
@@ -1,0 +1,4 @@
+  const checkUserSigned = useSignedCheck();
+
+  return () => {
+    if (!checkUserSigned()) return;
